+++ Source/CesiumRuntime/Private/Cesium3DTileset.cpp
@@ -134,7 +134,12 @@ ACesium3DTileset::ACesium3DTileset()
   PlatformName = UGameplayStatics::GetPlatformName();
 }
 
-ACesium3DTileset::~ACesium3DTileset() { this->DestroyTileset(); }
+ACesium3DTileset::~ACesium3DTileset() {
+  this->DestroyTileset();
+#if WITH_EDITOR
+  FEditorDelegates::PreBeginPIE.RemoveAll(this);
+#endif
+}
 
 PRAGMA_ENABLE_DEPRECATION_WARNINGS
 
@@ -833,6 +838,12 @@ void ACesium3DTileset::OnConstruction(const FTransform& Transform) {
   this->ResolveGeoreference();
   this->ResolveCameraManager();
   this->ResolveCreditSystem();
+
+#if WITH_EDITOR
+  FEditorDelegates::PreBeginPIE.AddUObject(
+      this,
+      &ACesium3DTileset::OnPreBeginPIE);
+#endif
 
   this->LoadTileset();
 
@@ -2097,12 +2108,12 @@ void ACesium3DTileset::Tick(float DeltaTime) {
   this->ResolveCameraManager();
   this->ResolveCreditSystem();
 
-  UCesium3DTilesetRoot* pRoot = Cast<UCesium3DTilesetRoot>(this->RootComponent);
-  if (!pRoot) {
+  if (this->SuspendUpdate) {
     return;
   }
 
-  if (this->SuspendUpdate) {
+  UCesium3DTilesetRoot* pRoot = Cast<UCesium3DTilesetRoot>(this->RootComponent);
+  if (!pRoot) {
     return;
   }
 
@@ -2433,6 +2444,12 @@ void ACesium3DTileset::RuntimeSettingsChanged(
           ->EnableExperimentalOcclusionCullingFeature;
   if (occlusionCullingAvailable != this->CanEnableOcclusionCulling) {
     this->CanEnableOcclusionCulling = occlusionCullingAvailable;
+    this->DestroyTileset();
+  }
+}
+
+void ACesium3DTileset::OnPreBeginPIE(bool bIsSimulating) {
+  if (this->UnloadEditorTilesInPlayMode) {
     this->DestroyTileset();
   }
 }